environ/requirements.txt
peewee==3.13.3
python-dotenv==0.14.0